+++ example/example.less
@@ -1,5 +1,5 @@
 #root,body,html{
-    height:100%
+    height:150%
 }
 body,html{
     margin:0;
@@ -7,7 +7,7 @@ body,html{
 }
 body{
     background: #C8C8A9;
-    overflow:hidden;
+    // overflow:hidden;
 }
 h2.example-title{
     padding: 40px;
