+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-jinke-music-player",
-  "version": "3.3.2",
+  "version": "3.4.0",
   "description": "Maybe the best beautiful HTML5 responsive player component for react",
   "main": "lib/index.js",
   "scripts": {
@@ -14,10 +14,10 @@
     "build:demo": "webpack --env.mode=PROD --progress --config  ./example/webpack.config.js",
     "demo": "webpack-dev-server --progress --inline --hot --config  ./example/webpack.config.js",
     "test": "jest __tests__/tests",
-    "coverage":"yarn test -- --coverage",
+    "coverage": "yarn test -- --coverage",
     "ci": "yarn coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "precommit": "lint-staged",
-    "prepare": "yarn build"
+    "prepublish": "yarn build"
   },
   "pre-commit": "lint",
   "lint-staged": {
@@ -61,26 +61,27 @@
     "react-icons": "^2.2.5"
   },
   "devDependencies": {
+    "@babel/cli": "7.0.0-beta.44",
+    "@babel/core": "7.0.0-beta.44",
+    "@babel/plugin-proposal-class-properties": "7.0.0-beta.44",
+    "@babel/plugin-proposal-object-rest-spread": "7.0.0-beta.44",
+    "@babel/plugin-syntax-dynamic-import": "7.0.0-beta.44",
+    "@babel/plugin-syntax-object-rest-spread": "7.0.0-beta.44",
+    "@babel/plugin-transform-object-assign": "7.0.0-beta.44",
+    "@babel/plugin-transform-runtime": "7.0.0-beta.44",
+    "@babel/preset-env": "7.0.0-beta.44",
+    "@babel/preset-react": "7.0.0-beta.44",
+    "@babel/preset-stage-0": "7.0.0-beta.44",
+    "@babel/preset-stage-1": "7.0.0-beta.44",
+    "@babel/runtime": "7.0.0-beta.44",
     "autoprefixer": "^6.7.2",
-    "babel-cli": "^6.16.0",
-    "babel-core": "6.x",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^8.2.3",
     "babel-jest": "^22.4.3",
-    "babel-loader": "6.x",
+    "babel-loader": "^8.0.0-beta.0",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-dynamic-import-node": "^1.0.2",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-syntax-object-rest-spread": "^6.13.0",
-    "babel-plugin-transform-class-properties": "^6.23.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "6.x",
-    "babel-preset-stage-1": "^6.24.1",
-    "babel-runtime": "^6.23.0",
     "chai": "^4.1.2",
     "copy-webpack-plugin": "latest",
     "coveralls": "^3.0.1",
@@ -122,6 +123,7 @@
     "regenerator-runtime": "^0.11.0",
     "rimraf": "^2.6.0",
     "style-loader": "~0.13.0",
+    "sweetalert": "^2.1.0",
     "url-loader": "^0.5.8",
     "webpack": "^2.7.0",
     "webpack-dev-server": "^2.3.0"
@@ -135,6 +137,9 @@
     "transformIgnorePatterns": [
       "/node_modules/",
       ".history/*"
+    ],
+    "modulePathIgnorePatterns": [
+      "/.history/"
     ],
     "moduleDirectories": [
       "node_modules",
