+++ src/index.js
@@ -1,5 +1,5 @@
 /**
- * @version 3.3.2
+ * @version 3.4.0
  * @name react-jinke-music-player
  * @description Maybe the best beautiful HTML5 responsive player component for react :)
  * @author Jinke.Li <1359518268@qq.com>
@@ -11,8 +11,9 @@ import classNames from "classnames";
 import isMobile from "is-mobile";
 import Slider from "rc-slider/lib/Slider";
 import Switch from "rc-switch";
-import AudioListsPanel from "./audioListsPanel";
-import AudioPlayerMobile from "./playerMobile";
+import {formatTime,createRandomNum} from "./utils"
+import AudioListsPanel from "./components/AudioListsPanel";
+import AudioPlayerMobile from "./components/PlayerMobile";
 
 import FaHeadphones from "react-icons/lib/fa/headphones";
 import FaMinusSquareO from "react-icons/lib/fa/minus-square-o";
@@ -61,27 +62,7 @@ const PlayModel = ({ visible, value }) => (
   </div>
 );
 
-//秒转换成 时间格式
-export function formatTime(second) {
-  let i = 0;
-  let s = parseInt(second);
-  if (s > 60) {
-    i = parseInt(s / 60);
-    s = parseInt(s % 60);
-    if (i > 60) {
-      i = parseInt(i % 60);
-    }
-  }
-  // 补零
-  const zero = v => (v >> 0 < 10 ? `0${v}` : v);
-  return [zero(i), zero(s)].join(":");
-}
-
-export function createRandomNum(minNum, maxNum) {
-  return parseInt(Math.random() * (maxNum - minNum + 1) + minNum, 10);
-}
-
-export const sliderBaseOptions = {
+const sliderBaseOptions = {
   min: 0,
   step: 0.01,
   trackStyle: { backgroundColor: "#31c27c" },
@@ -502,12 +483,12 @@ export default class ReactJkMusicPlayer extends PureComponent {
             <audio
               key="audio"
               className="music-player-audio"
-              preload="auto"
               src={musicSrc}
               ref={node => (this.audio = node)}
             />
           </div>
-          {toggle ? (
+        </div>
+        {toggle ? (
             isMobile ? (
               undefined
             ) : (
@@ -691,7 +672,6 @@ export default class ReactJkMusicPlayer extends PureComponent {
           ) : (
             undefined
           )}
-        </div>
         {/* 播放列表面板 */}
         <AudioListsPanel
           playId={playId}
