+++ src/styles/playerMobile.less
@@ -1,0 +1,202 @@
+@import "./vars.less";
+@import "./mixins.less";
+.react-jinke-music-player-mobile {
+    @common-animate-time: .45s cubic-bezier(0.165, 0.84, 0.44, 1.1);
+    height: 100vh;
+    width: 100%;
+    position: absolute;
+    z-index: 999;
+    left: 0;
+    top: 0;
+    background-color: rgba(0, 0, 0, .6);
+    padding: 20px;
+    color: #fff;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    animation: mobile-bg @common-animate-time forwards;
+    transition: background-color @common-animate-time;
+    >.group {
+        flex: 1 1 auto;
+    }
+    svg {
+        transition: color @common-animate-time;
+    }
+    &-play-model-tip {
+        position: absolute;
+        left: 0;
+        width: 100%;
+        top: 0;
+        height: 35px;
+        line-height: 35px;
+        background-color: @primary-color;
+        color: #fff;
+        padding: 0 20px;
+        z-index: 1000;
+        box-shadow: 0 2px 20px rgba(0, 0, 0, .1);
+        transform: translate3d(0, -105%, 0);
+        .transition();
+        .title {
+            margin-right: 12px;
+        }
+        .text {
+            font-size: 14px;
+        }
+        &.show {
+            transform: translate3d(0, 0, 0);
+        }
+    }
+    &-header {
+        width: 100%;
+        top: 0;
+        left: 0;
+        display: flex;
+        align-items: center;
+        animation: fromDown @common-animate-time forwards;
+        .title {
+            flex-basis: 80%;
+            text-align: center;
+            font-size: 20px;
+            .ellipsis-1()
+        }
+        .item {
+            display: inline-flex;
+            width: 50px;
+        }
+        .right {
+            justify-content: flex-end;
+            color: @base-color;
+            svg {
+                font-size: 25px;
+            }
+        }
+    }
+    &-singer {
+        animation: fromDown @common-animate-time forwards;
+        padding: 12px 0;
+        @media screen and (max-width: 320px) {
+            padding: 0;
+        }
+        .name {
+            position: relative;
+            font-size: 14px;
+            &:before,
+            &:after {
+                content: "";
+                position: absolute;
+                width: 16px;
+                height: 1px;
+                top: 9px;
+                border-radius: 2px;
+                background-color: @base-color;
+            }
+            &:after {
+                left: -25px;
+            }
+            &:before {
+                right: -25px;
+            }
+        }
+    }
+    &-cover {
+        width: 300px;
+        height: 300px;
+        border-radius: 50%;
+        border: 5px solid rgba(0, 0, 0, .2);
+        box-shadow: 0 0 1px 3px rgba(0, 0, 0, .1);
+        margin: 15px auto;
+        animation: fromTo @common-animate-time forwards;
+        overflow: hidden;
+        @media screen and (max-width: 320px) {
+            width:230px;
+            height: 230px;
+            margin: 10px auto;
+        }
+        .cover {
+            width: 100%;
+            animation: imgRotate 30s linear infinite;
+        }
+    }
+    &-progress {
+        justify-content: space-around;
+        display: flex;
+        align-items: center;
+        .current-time,
+        .duration {
+            display: inline-flex;
+            width: 55px;
+            font-size: 12px;
+            color: @base-color;
+        }
+        .duration {
+            justify-content: flex-end;
+        }
+        .progress-bar {
+            flex: 1 1 auto;
+        }
+        .rc-slider-rail {
+            background-color: @base-color;
+        }
+    }
+    &-switch {
+        animation: fromDown @common-animate-time forwards;
+    }
+    &-toggle {
+        padding: 17px 0;
+        .group {
+            svg {
+                font-size: 40px;
+            }
+        }
+        .play-btn {
+            padding: 0 40px;
+            svg {
+                font-size: 60px;
+            }
+        }
+        @media screen and (max-width:320px) {
+            padding: 10px 0;
+            >.group {
+                svg {
+                    font-size: 32px;
+                }
+            }
+            .play-btn {
+                svg {
+                    font-size: 50px;
+                }
+            }
+        }
+    }
+    &-toggle,
+    &-progress {
+        animation: fromTo @common-animate-time forwards;
+    }
+    &-operation {
+        animation: fromTo @common-animate-time forwards;
+        .items {
+            display: flex;
+            justify-content: space-around;
+            align-items: center;
+            .item {
+                flex: 1;
+                text-align: center;
+                svg {
+                    color: @base-color;
+                    font-size: 25px;
+                    @media screen and (max-width: 320px) {
+                        font-size:20px;
+                    }
+                }
+            }
+        }
+    }
+    @keyframes mobile-bg {
+        from {
+            opacity: 0
+        }
+        to {
+            opacity: 1;
+        }
+    }
+}
