+++ src/styles/vars.less
@@ -16,3 +16,5 @@
 @player-width:80px;
 @player-height:@player-width;
 @base-color:rgba(255,255,255,.6);
+@common-animate-type:"cubic-bezier(0.25, 0.46, 0.33, 0.98)";
+@common-animate-time:"400 ms cubic-bezier(0.25, 0.46, 0.33, 0.98)";
